extracted metadataresponse class out into its own file
